refactor(spectrogram): move computation + worker into @dawcore/spectrogram
Relocate the pure FFT computation, Web Worker, and fft.js type shim from
@waveform-playlist/spectrogram into the new @dawcore/spectrogram package.
This positions the framework-agnostic primitives for sharing between the
React Provider (later commit) and the upcoming dawcore Lit element.

- 12 source files + 6 test files moved via git mv (full history preserved)
- @dawcore/spectrogram/src/index.ts now exports computation + worker
- tsconfig drops rootDir so __tests__/ at package root coexists with src/
- tsup.config enables the worker entry now that the file is present

Slim @waveform-playlist/spectrogram is left transiently broken; the next
commit repoints its imports and bumps the version.
