+++ packages/dawcore-spectrogram/src/computation/colorMaps.ts
