+++ packages/dawcore-spectrogram/src/computation/computeSpectrogram.ts
