+++ packages/dawcore-spectrogram/src/computation/frequencyScales.ts
