+++ packages/dawcore-spectrogram/src/computation/index.ts
