+++ packages/dawcore-spectrogram/src/computation/windowFunctions.ts
