+++ packages/dawcore-spectrogram/src/types/fft.js.d.ts
