+++ packages/dawcore-spectrogram/src/worker/createSpectrogramWorkerPool.ts
