+++ packages/dawcore-spectrogram/src/worker/index.ts
