+++ packages/dawcore-spectrogram/src/worker/spectrogram.worker.ts
