+++ packages/dawcore-spectrogram/tsup.config.ts
@@ -12,17 +12,16 @@ export default defineConfig([
     sourcemap: true,
     clean: true,
   },
-  // Web Worker — ESM only, no DTS, don't clean (would wipe first build's output)
-  // NOTE: Uncomment when src/worker/spectrogram.worker.ts and src/orchestrator/index.ts are created (Task 2, Task 7)
-  // {
-  //   entry: {
-  //     'worker/spectrogram.worker': 'src/worker/spectrogram.worker.ts',
-  //     'orchestrator/index': 'src/orchestrator/index.ts',
-  //   },
-  //   format: ['esm'],
-  //   dts: false,
-  //   splitting: false,
-  //   sourcemap: true,
-  //   clean: false,
-  // },
+  // Web Worker — ESM only, no DTS, don't clean (would wipe first build's output).
+  // The orchestrator entry is added in Task 7 when src/orchestrator/index.ts exists.
+  {
+    entry: {
+      'worker/spectrogram.worker': 'src/worker/spectrogram.worker.ts',
+    },
+    format: ['esm'],
+    dts: false,
+    splitting: false,
+    sourcemap: true,
+    clean: false,
+  },
 ]);
